+++ packages/auth/__tests__/providers/cognito/signInStateManagement.test.ts
@@ -3,20 +3,11 @@
 
 import { Amplify } from '@aws-amplify/core';
 
-import {
-	confirmSignIn,
-	getCurrentUser,
-	signIn,
-} from '../../../src/providers/cognito';
+import { getCurrentUser, signIn } from '../../../src/providers/cognito';
 import * as signInHelpers from '../../../src/providers/cognito/utils/signInHelpers';
-import * as clients from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider';
 import { signInStore } from '../../../src/providers/cognito/utils/signInStore';
-import {
-	AssociateSoftwareTokenCommandOutput,
-	RespondToAuthChallengeCommandOutput,
-} from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider/types';
+import { RespondToAuthChallengeCommandOutput } from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider/types';
 import { cognitoUserPoolsTokenProvider } from '../../../src/providers/cognito/tokenProvider';
-import { mfaSetupStore } from '../../../src/providers/cognito/utils/mfaSetupStore';
 
 import { authAPITestParams } from './testUtils/authApiTestParams';
 
@@ -118,192 +109,5 @@ describe('local sign-in state management tests', () => {
 
 		handleUserSRPAuthflowSpy.mockClear();
 		mockedGetCurrentUser.mockClear();
-	});
-});
-
-describe('mfa setup state management tests', () => {
-	const handleChallengeNameSpy = jest.spyOn(
-		signInHelpers,
-		'handleChallengeName',
-	);
-	const handleUserSRPAuthFlowSpy = jest.spyOn(
-		signInHelpers,
-		'handleUserSRPAuthFlow',
-	);
-
-	beforeAll(() => {
-		Amplify.configure({ Auth: authConfig });
-	});
-
-	test('mfa setup state should initialize as undefined', () => {
-		const mfaSetupState = mfaSetupStore.getState();
-
-		expect(mfaSetupState).toStrictEqual(undefined);
-	});
-
-	test('mfa setup state should be updated with options during first MFA_SETUP challenge', async () => {
-		jest.spyOn(signInHelpers, 'handleUserSRPAuthFlow').mockImplementationOnce(
-			async (): Promise<RespondToAuthChallengeCommandOutput> => ({
-				ChallengeName: 'MFA_SETUP',
-				Session: '1234234232',
-				$metadata: {},
-				ChallengeParameters: {
-					MFAS_CAN_SETUP: '["SMS_MFA","SOFTWARE_TOKEN_MFA", "EMAIL_OTP"]',
-				},
-			}),
-		);
-
-		await signIn({ username, password });
-
-		const mfaSetupState = mfaSetupStore.getState();
-
-		expect(mfaSetupState).toStrictEqual({
-			status: 'IN_PROGRESS',
-			options: ['TOTP', 'EMAIL'],
-		});
-	});
-
-	test('mfa setup state should be updated with selected value during second MFA_SETUP challenge', async () => {
-		jest.spyOn(signInHelpers, 'handleUserSRPAuthFlow').mockImplementationOnce(
-			async (): Promise<RespondToAuthChallengeCommandOutput> => ({
-				ChallengeName: 'MFA_SETUP',
-				Session: '1234234232',
-				$metadata: {},
-				ChallengeParameters: {
-					MFAS_CAN_SETUP: '["SMS_MFA","SOFTWARE_TOKEN_MFA", "EMAIL_OTP"]',
-				},
-			}),
-		);
-
-		await signIn({ username, password });
-
-		await confirmSignIn({
-			challengeResponse: 'EMAIL',
-		});
-
-		expect(mfaSetupStore.getState()).toStrictEqual({
-			status: 'COMPLETE',
-			options: ['TOTP', 'EMAIL'],
-			value: 'EMAIL',
-		});
-	});
-
-	test('mfa setup state should be cleared with successful sign in', async () => {
-		handleUserSRPAuthFlowSpy.mockImplementationOnce(
-			async (): Promise<RespondToAuthChallengeCommandOutput> =>
-				authAPITestParams.RespondToAuthChallengeMultipleMfaSetupOutput,
-		);
-
-		await signIn({ username, password });
-
-		await confirmSignIn({
-			challengeResponse: 'EMAIL',
-		});
-
-		handleChallengeNameSpy.mockImplementationOnce(
-			async (): Promise<RespondToAuthChallengeCommandOutput> => ({
-				ChallengeName: 'EMAIL_OTP',
-				Session: '1234234232',
-				$metadata: {},
-				ChallengeParameters: {
-					CODE_DELIVERY_DELIVERY_MEDIUM: 'EMAIL',
-					CODE_DELIVERY_DESTINATION: 'j***@a***',
-				},
-			}),
-		);
-
-		await confirmSignIn({
-			challengeResponse: 'j***@a***',
-		});
-
-		handleChallengeNameSpy.mockImplementationOnce(
-			async (): Promise<RespondToAuthChallengeCommandOutput> =>
-				authAPITestParams.RespondToAuthChallengeCommandOutput,
-		);
-
-		await confirmSignIn({
-			challengeResponse: '123456',
-		});
-
-		const mfaSetupState = mfaSetupStore.getState();
-
-		expect(mfaSetupState).toStrictEqual(undefined);
-	});
-
-	test('mfa setup state should be reset with each sign in attempt', async () => {
-		handleUserSRPAuthFlowSpy.mockImplementationOnce(
-			async (): Promise<RespondToAuthChallengeCommandOutput> =>
-				authAPITestParams.RespondToAuthChallengeMultipleMfaSetupOutput,
-		);
-
-		await signIn({ username, password });
-
-		await confirmSignIn({
-			challengeResponse: 'EMAIL',
-		});
-
-		expect(mfaSetupStore.getState()).toStrictEqual({
-			status: 'COMPLETE',
-			options: ['TOTP', 'EMAIL'],
-			value: 'EMAIL',
-		});
-
-		handleUserSRPAuthFlowSpy.mockImplementationOnce(
-			async (): Promise<RespondToAuthChallengeCommandOutput> =>
-				authAPITestParams.RespondToAuthChallengeCommandOutput,
-		);
-
-		await signIn({ username, password });
-
-		expect(mfaSetupStore.getState()).toStrictEqual(undefined);
-	});
-
-	test('mfa setup state should autocomplete when only one allowed MFA setup option is available (EMAIL_OTP)', async () => {
-		handleUserSRPAuthFlowSpy.mockImplementationOnce(
-			async (): Promise<RespondToAuthChallengeCommandOutput> => ({
-				ChallengeName: 'MFA_SETUP',
-				Session: '1234234232',
-				$metadata: {},
-				ChallengeParameters: {
-					MFAS_CAN_SETUP: '["SMS_MFA", "EMAIL_OTP"]',
-				},
-			}),
-		);
-
-		await signIn({ username, password });
-
-		expect(mfaSetupStore.getState()).toStrictEqual({
-			status: 'COMPLETE',
-			options: ['EMAIL'],
-			value: 'EMAIL',
-		});
-	});
-	test('mfa setup state should autocomplete when only one allowed MFA setup option is available (SOFTWARE_TOKEN_MFA)', async () => {
-		handleUserSRPAuthFlowSpy.mockImplementationOnce(
-			async (): Promise<RespondToAuthChallengeCommandOutput> => ({
-				ChallengeName: 'MFA_SETUP',
-				Session: '1234234232',
-				$metadata: {},
-				ChallengeParameters: {
-					MFAS_CAN_SETUP: '["SMS_MFA", "SOFTWARE_TOKEN_MFA"]',
-				},
-			}),
-		);
-
-		jest.spyOn(clients, 'associateSoftwareToken').mockImplementationOnce(
-			async (): Promise<AssociateSoftwareTokenCommandOutput> => ({
-				SecretCode: 'secret-code',
-				Session: '12341234',
-				$metadata: {},
-			}),
-		);
-
-		await signIn({ username, password });
-
-		expect(mfaSetupStore.getState()).toStrictEqual({
-			status: 'COMPLETE',
-			options: ['TOTP'],
-			value: 'TOTP',
-		});
 	});
 });
