+++ packages/auth/src/providers/cognito/apis/signInWithCustomAuth.ts
@@ -32,7 +32,6 @@ import {
 } from '../utils/clients/CognitoIdentityProvider/types';
 import { tokenOrchestrator } from '../tokenProvider';
 import { dispatchSignedInHubEvent } from '../utils/dispatchSignedInHubEvent';
-import { resetMfaSetupState } from '../utils/mfaSetupStore';
 
 /**
  * Signs a user in using a custom authentication flow without password
@@ -65,7 +64,6 @@ export async function signInWithCustomAuth(
 	);
 
 	try {
-		resetMfaSetupState();
 		const {
 			ChallengeName: retriedChallengeName,
 			ChallengeParameters: retiredChallengeParameters,
