+++ packages/auth/src/providers/cognito/apis/signInWithUserPassword.ts
@@ -32,7 +32,6 @@ import {
 import { cacheCognitoTokens } from '../tokenProvider/cacheTokens';
 import { tokenOrchestrator } from '../tokenProvider';
 import { dispatchSignedInHubEvent } from '../utils/dispatchSignedInHubEvent';
-import { resetMfaSetupState } from '../utils/mfaSetupStore';
 
 /**
  * Signs a user in using USER_PASSWORD_AUTH AuthFlowType
@@ -65,7 +64,6 @@ export async function signInWithUserPassword(
 	);
 
 	try {
-		resetMfaSetupState();
 		const {
 			ChallengeName: retiredChallengeName,
 			ChallengeParameters: retriedChallengeParameters,
