+++ packages/auth/src/providers/cognito/utils/types.ts
@@ -140,10 +140,3 @@ function isAuthenticatedWithImplicitOauthFlow(
 ) {
 	return isAuthenticated(tokens) && !tokens?.refreshToken;
 }
-
-export type Store<State, Action> = (reducer: Reducer<State, Action>) => {
-	getState(): ReturnType<Reducer<State, Action>>;
-	dispatch(action: Action): void;
-};
-
-export type Reducer<State, Action> = (state: State, action: Action) => State;
